Zadatak 1. If / Scanner / scanner.close() added. int sum changed to double.

+++ src/Inputs/Zadatak1.java
@@ -17,7 +17,7 @@ public class Zadatak1 {
 
         scan.close();
 
-        int sum = a + b;
+        double sum = a + b;
 
         double m = (sum / 2);
